I think I made it follow the specific rules
I was very confused on this one, but after a while I realized what I
needed to do in regards with the selectors I think I got most of them
correct. After looking at the solutions to check my answers, I fixed
some of the selectors and also took note of the fact that some of them
were a lot better ideas than mine.