Finished assignment 02 in the css-exercises
This one took me quite a bit longer to figure out, and I dont think I
actually fully figured it out until I looked at the solutions and put it
into practice in my own stylesheet. I now understand that you have to be
careful with how you setup the margins and padding for the box model and
normal flow.